+++ apps/sdk/src/indexSSR.js
@@ -1,9 +1,6 @@
 // Packages
 import { renderToPipeableStream } from 'react-dom/server';
 
-// Monorepo
-import { init, shared } from '@plitzi/sdk-elements/PluginRemote';
-
 // Relatives
 import PlitziSdk, {
   ComponentProvider,
@@ -27,9 +24,7 @@ export {
   RENDER_MODE_RAW,
   RENDER_MODE_SHADOW,
   RENDER_MODE_SSR,
-  renderToPipeableStream,
-  init,
-  shared
+  renderToPipeableStream
 };
 
 export default PlitziSdk;