+++ apps/sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @plitzi/plitzi-sdk
 
+## 0.22.17
+
+### Patch Changes
+
+- v0.22.17
+- Updated dependencies
+  - @plitzi/sdk-elements@0.22.17
+  - @plitzi/sdk-shared@0.22.17
+  - @plitzi/sdk-auth@0.22.17
+  - @plitzi/sdk-dev-tools@0.22.17
+  - @plitzi/sdk-event-bridge@0.22.17
+  - @plitzi/sdk-interactions@0.22.17
+  - @plitzi/sdk-navigation@0.22.17
+  - @plitzi/sdk-plugins@0.22.17
+  - @plitzi/sdk-state@0.22.17
+
 ## 0.22.16
 
 ### Patch Changes