+++ packages/sdk-data-source/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @plitzi/sdk-data-source
 
+## 0.22.17
+
+### Patch Changes
+
+- v0.22.17
+- Updated dependencies
+  - @plitzi/sdk-shared@0.22.17
+  - @plitzi/sdk-event-bridge@0.22.17
+
 ## 0.22.16
 
 ### Patch Changes