+++ packages/sdk-interactions/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @plitzi/sdk-interactions
 
+## 0.22.17
+
+### Patch Changes
+
+- v0.22.17
+- Updated dependencies
+  - @plitzi/sdk-shared@0.22.17
+
 ## 0.22.16
 
 ### Patch Changes